Move state_change to a new module
